changes after code review comments
